pyproject.toml
[project]
name = "asr-nlu-pipeline"
version = "0.1.0"
requires-python = ">=3.13,<3.14"
dependencies = [
  "dotenv>=0.9.9",
  "gradio==6.2.0",
  "nemo-toolkit[asr]==2.6.0",
  "ollama>=0.6.1",
  "onnx<1.19.0",
  "seqeval>=1.2.2",
  "soundfile>=0.13.1",
  "torch==2.9.1",
  "torchvision==0.24.1",
  "typer==0.20.1",
  "wordninja>=2.0.0",
]

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[tool.uv.sources]
torch = [
  { index = "pytorch-cu128", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchvision = [
  { index = "pytorch-cu128", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
